Implement String.Chars protocol for Schema types
This might be a bad choice, but for #53 I need a convenient way to get
the name of the type. I thought I could use `to_string(%Enum{})` for
example, to get the type and display it in the error message.

Counter argument for doing this is that we might want to show a fuller
representation of the type. Thoughts?

- [x] Primitive
- [x] Enum
- [x] Fixed
- [ ] Array
- [ ] Map
- [ ] Record
- [ ] Union